refactor  cartcontroler and cartitemcontroller; add @CrossOrigin ; check before delete product ; change url : admin/api/cart and client/api/cartitems.

+++ src/main/java/com/javaBTL/BE_JAVA_BTL/service/ProductServiceImpl.java
@@ -4,15 +4,17 @@ import com.javaBTL.BE_JAVA_BTL.model.Product;
 import com.javaBTL.BE_JAVA_BTL.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.CrossOrigin;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
 public class ProductServiceImpl implements ProductService {
     @Autowired
     private ProductRepository productRepository;
-
+    @CrossOrigin
     @Override
     public Product createProduct(Product product) {
         productRepository.save(product);
@@ -20,6 +22,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
+    @CrossOrigin
     public Product createOrUpdateProduct(UUID id, Product product) {
         if (productRepository.existsById(id)) {
             // Cập nhật sản phẩm theo id
@@ -30,6 +33,7 @@ public class ProductServiceImpl implements ProductService {
             return productRepository.save(product);
         }
     }
+    @CrossOrigin
     @Override
     public List<Product> uploadProducts(List<Product> products) {
         // Iterate through the list of products and save each one
@@ -46,40 +50,55 @@ public class ProductServiceImpl implements ProductService {
         }
         return products;
     }
+    @CrossOrigin
     @Override
     public List<Product> getAllProducts() {
         return productRepository.findAll();
     }
+    @CrossOrigin
 
     @Override
     public List<Product> getAllLaptops() {
         return productRepository.findProductsByCategory("laptop");
     }
+    @CrossOrigin
 
     @Override
     public Product getProductById(UUID id) {
         return productRepository.findById(id).orElse(null);
     }
+    @CrossOrigin
 
     @Override
     public List<Product> searchProducts(String keyword) {
         // Triển khai logic tìm kiếm sản phẩm dựa trên từ khóa ở đây và trả về danh sách sản phẩm tương ứng
-        return productRepository.searchProducts(keyword);
+       if(productRepository.searchProducts(keyword)  != null) {
+           return productRepository.searchProducts(keyword);
+       } else {
+           return null;
+       }
     }
 
-
+    @CrossOrigin
     @Override
-    public void deleteProduct(UUID id) {
-        // Triển khai logic xóa sản phẩm theo ID ở đây
-        productRepository.deleteById(id);
+    public boolean deleteProduct(UUID id) {
+        Optional<Product> product = productRepository.findById(id);
+        if(product.isPresent()) {
+            productRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
     public List<Product> getProductsByCategory(String category){
         return productRepository.getProductsByCategory(category);
     }
+    @CrossOrigin
     @Override
     public void deleteAllProducts() {
         productRepository.deleteAll();
     }
+    @CrossOrigin
     @Override
     public void deleteProductsByCategory(String category) {
         productRepository.deleteProductsByCategory(category);
